// src/app/api/models/authentication.model.ts
export declare type TaiKhoanModel = {
  account_id:string,
  id?: string;
  tenhienthi?: string;
  username?: string;
  password?: string;
  laplaimatkhau?: string;
  email?: string;
  avatar?: string;
  ngaytao?: string;
  trangthai?: string;
  roleid?:number;
  role_id?:number;
  namerole?:string;
  role_name?:string;
  first_name?:string;
  last_name?:string;
  phone_number?:string;
  user_id?:number;
};